- fixed several warnings regarding signed/unsigned ops, incomplet   initializers for static data.
  Based on a patch send by Eliot Gable <egable@broadvox.net> for increasing
  the level of warning messages




git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1782 689a6050-402a-0410-94f2-e92a70836424

+++ action.c
@@ -507,7 +507,7 @@ int do_action(struct action* a, struct sip_msg* msg)
 				ret=E_BUG;
 				break;
 			}
-			ret = msg->len >= a->elem[0].u.number ? 1 : -1;
+			ret = (msg->len >= (unsigned int)a->elem[0].u.number) ? 1 : -1;
 			break;
 		case SETFLAG_T:
 			ret = setflag( msg, a->elem[0].u.number );
